Move download ffi function to download module
